requirements.txt
fastapi
gunicorn
uvicorn
numpy
requests
transformers
sentencepiece
tritonclient[all]
onnx
onnxruntime-gpu
sympy
coloredlogs
pycuda
nvidia-pyindex
pytest
black
isort
flake8
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.0+cu113
nvidia-tensorrt==8.0.3.4
onnx_graphsurgeon
polygraphy==0.33.0
triton-model-analyzer
